Rename the package to msentraid
The linter is now linting the ex microsoft_entra_id package and is
unhappy about the namng.
