+++ internal/providers/msentraid/msentraid.go
@@ -1,5 +1,5 @@
-// Package microsoft_entra_id is the Microsoft Entra ID specific extension.
-package microsoft_entra_id
+// Package msentraid is the Microsoft Entra ID specific extension.
+package msentraid
 
 import (
 	"context"
